+++ energy/utils/globals.py
@@ -6,6 +6,7 @@
 
 KEY_CURRENTCOST = "currentcost"
 KEY_ENERGY = "energy"
+ERROR_NO_CC_SETTINGS = "currentcost_no_db_settings"
 ERROR_CC_BAD_PORT = "currentcost_bad_port"
 ERROR_CC_NO_MESSAGE = "currentcost_no_message"
 ERROR_CC_DISCONNECTED = "currentcost_disconnected"
@@ -13,6 +14,7 @@ ERROR_CC_INCORRECT_MESSAGE = "currentcost_incorrect_message"
 ERROR_CC_INCORRECT_MESSAGE_MISSING_TMPR = "currentcost_incorrect_message_missing_tmpr"
 ERROR_CC_INCORRECT_MESSAGE_MISSING_WATTS = "currentcost_incorrect_message_missing_watts"
 ERROR_CURRENTCOST = {
+    ERROR_NO_CC_SETTINGS: "There is no CurrentCost settings in DB",
     ERROR_CC_BAD_PORT: "CurrentCost %s in %s: TTY problem: %s is unreachable. Retry connection in %s seconds.",
     ERROR_CC_NO_MESSAGE: "CurrentCost %s in %s: Reach timeout. Verify CurrentCost wire connection or wave range",
     ERROR_CC_DISCONNECTED: "CurrentCost %s in %s: TTY port %s disconnected.",